Rename `channel` module to `channels`
